App/requirements.txt
pandas
torch
tensorflow
tf-keras
streamlit
SpeechRecognition
moviepy
pytube
ffmpeg-python
pydub
beautifulsoup4
transformers
llama-index-vector-stores-lancedb
llama-index-multi-modal-llms-openai
llama-index-multi-modal-llms-openai
llama-index-vector-stores-lancedb
llama-index-embeddings-clip
llama_index 
ftfy 
regex 
tqdm
openai-whisper
torchvision
matplotlib 
scikit-image
lancedb
soundfile
llama-index-multi-modal-llms-gemini
llama-index-vector-stores-qdrant
llama-index-embeddings-gemini
llama-index-llms-gemini
llama-index
google-generativeai>=0.3.0
matplotlib
qdrant_client
transformers
peft
accelerate
optimum
langchain
einops
rouge_score
bert_score
pytube
ffmpeg-python
SpeechRecognition
pydub
moviepy
openai
python-dotenv
scikit-learn
accelerate
ipython
